refactor: src/js/eventListener.js의 경로를 src/js/evnet/idnex.js로 수정
